Add unit tests
X86 specific test has been disabled as it fails when unbundling bochs.
This needs to be looked into, as either it's a false positive and tests
need fixes, or if there's a bug in the unbundling logic.

Gtest code is bundled. It should be evaluated how (if possible) to
unbundle this.