Update to 5.0.13669
This bundles glslang. It seems there were some changes that is making
this hard to maintain. I've kept the unbundled state of previous versions,
partially due to a misunderstanding and to keep things as-is.

To reiterate, this is a static library, which dolphin updates often on it's
own cadence. It seems of little value to always patch dolphin to allign
with fedora's version. This is a maintainence burden with no value to the
users, and little value to the distro if dolphin is kept up to date.

As well, even though it uses more build time, it uses less space since
dolphin optimizes the library a bit better, since it doesn't need all
functionality. I.e. build time impact seems low, with less space use.