Bundle glslang
As much as bundling is ideal, it seems that glslang is too unstable to
reliably use as a build dependency. API breakage within a release branch
seems to be the norm and makes maintainace of this package hard.

Furthermore, the library is static by nature, so it demotes the value of
having it unbundled in my eyes, as the only thing it saves is build time,
having no affect on the end user, nor space usage.

Furthermore, upstream does not care to allow building against a shared
static glslang, nor is there any API stability to allow for this in a
reasonable way.